Migrate from CMake/Catch to Bazel/GTest
