+++ MODULE.bazel
@@ -1,0 +1,13 @@
+"""N-Body Gravitational Simulation Software"""
+
+module(name = "markets", version = "1.0.0")
+
+bazel_dep(name = "rules_cc", version = "0.1.1")
+bazel_dep(name = "protobuf", version = "29.3")
+bazel_dep(name = "googletest", version = "1.15.2")
+bazel_dep(name = "hedron_compile_commands", dev_dependency = True)
+git_override(
+    module_name = "hedron_compile_commands",
+    remote = "https://github.com/hedronvision/bazel-compile-commands-extractor.git",
+    commit = "0e990032f3c5a866e72615cf67e5ce22186dcb97",
+)
