+++ nbsim/core/octree/octree_node_type.hpp
