+++ nbsim/core/vec3/vec3.cpp
@@ -1,0 +1,6 @@
+#include "nbsim/core/vec3/vec3.hpp"
+
+std::ostream &operator<<(std::ostream &os, const Vec3 &vec) {
+  os << "[" << vec.x << "," << vec.y << "," << vec.z << "]";
+  return os;
+}
