--- test/test.cpp
@@ -1,4 +1,0 @@
-#define CATCH_CONFIG_MAIN
-#include "catch.h"
-
-// A simple file that acts as an aggregate for all other test cases
